% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print_df}
\alias{knit_print_df}
\title{Automatically print data.frames as reactable html widget}
\usage{
knit_print_df(df, options, ...)
}
\arguments{
\item{df}{a \code{data.frame}}

\item{options}{this will be taken as a list of a parameters inputted into the
chunk. Specifically, the \code{fig.cap} chunk parameter is used to input to
label tables with a caption enabling referencing them in the text.}

\item{...}{Included based off this tutorial
https://cran.r-project.org/web/packages/knitr/vignettes/knit_print.html.
Required for \code{knitr::knit_print}.}
}
\value{
reactable html widget with a caption.
}
\description{
This is intended to replace the default \code{knitr} printing method for
\code{data.frame} objects and is used in the "rmd_template_dz" to automatically
render \code{data.frame}s as \code{reactable}s. This also adds a caption to enable the
a workaround to allow table cross-referencing in the .Rmd using
\verb{\\@ref(tab:KEY)}. The solution uses \code{htmltools} to manually add a caption
inputted through the chunk option \code{fig.cap} and was inspired from
https://github.com/glin/reactable/issues/15.
}
\examples{

\dontrun{
# add this line to the setup chunk of your .Rmd
# to overwrite the existing method for printing data.frames
registerS3method("knit_print", "data.frame", Rmdplate::knit_print.data.frame)
}
}
