% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print_reactable}
\alias{knit_print_reactable}
\title{Default printing method for \code{reactable}s}
\usage{
knit_print_reactable(df_react, options, ...)
}
\arguments{
\item{df_react}{\code{reactable} data table.}

\item{options}{this will be taken as a list of a parameters inputted into the
chunk. Specifically, \code{fig.cap} is used to}

\item{...}{Included based off this tutorial
https://cran.r-project.org/web/packages/knitr/vignettes/knit_print.html.
Required for \code{knitr::knit_print}.}
}
\value{
reactable html widget with a caption.
}
\description{
This is added to ensure that, say you convert the \code{data.frame} to a
\code{reactable} in a code chunk (therefore this will not be passed through
\code{knit_print.data.frame}), you can still use \code{fig.cap}. This adds a caption to
the \code{reactable} html widget.
}
