% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{list_template_dirs}
\alias{list_template_dirs}
\alias{plate}
\title{Interactively choose, knit, and show R Markdown Drafts.}
\usage{
list_template_dirs()
}
\value{
invisible()
}
\description{
This function provides non-RStudio users with alternative methods to interactively choose Rmd templates based upon already-existing packages on system, help users edit their Rmd drafts in a preferable text editor, quickly knit it and show the results simultaneously.
}
\details{
plate

You can use this function to conveniently draft your R Markdown documents based on other templates provided by the available packages installed on your system.
}
\examples{
if(interactive()) {
# Use rmdplate::plate(), then follow the instructions.
}
}
\author{
JooYoung Seo, \email{jooyoung@psu.edu}
}
\keyword{plate}
