% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all.R
\name{render_all}
\alias{render_all}
\title{Interactively choose, knit, and show R Markdown Drafts.}
\usage{
render_all(input, output_format = "all", ...)
}
\value{
invisible()
}
\description{
This function provides non-RStudio users with alternative methods to interactively choose Rmd temrender_alls based upon already-existing packages on system, help users edit their Rmd drafts in a preferable text editor, quickly knit it and show the results simultaneously.
}
\details{
render_all

You can use this function to conveniently draft your R Markdown documents based on other temrender_alls provided by the available packages installed on your system.
}
\examples{
if(interactive()) {
# Use rmdplate::render_all(), then follow the instructions.
}
}
\author{
JooYoung Seo, \email{jooyoung@psu.edu}
}
\keyword{render_all}
