% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdplate-package.R
\docType{package}
\name{rmdplate-package}
\alias{rmdplate}
\alias{rmdplate-package}
\title{rmdplate: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jooyoungseo/rmdplate}
  \item Report bugs at \url{https://github.com/jooyoungseo/rmdplate/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
